test(backend): split test_refactored_helpers.py into module-specific files
Rename and split `backend/tests/test_refactored_helpers.py` into `test_memory.py`, `test_openrouter.py`, and `test_crew.py` to match the modules they test.
